Integrate type union regressions into existing tests
