+++ tests/main/jsonschema/test_main_jsonschema.py
@@ -23537,3 +23537,150 @@ def test_compound_property_names_compatibility(
         invalid_json=json.dumps(payload["invalid"]),
         expected_error_type="dict_type",
     )
+
+
+TYPE_UNION_INPUTS = JSON_SCHEMA_DATA_PATH / "type_union_constraints"
+
+
+TYPE_UNION_PAYLOADS = DATA_PATH / "payloads" / "type_union_constraints"
+
+
+TYPE_UNION_EXPECTED = EXPECTED_JSON_SCHEMA_PATH / "type_union_constraints"
+
+
+TYPE_UNION_CASES = json.loads((TYPE_UNION_PAYLOADS / "cases.json").read_text())
+
+
+@pytest.mark.parametrize("name", TYPE_UNION_CASES)
+@pytest.mark.parametrize("constraints", [False, True])
+@pytest.mark.parametrize("entry", ["cli", "api", "dynamic"])
+def test_type_union_constraints(name: str, constraints: bool, entry: str, output_file: Path) -> None:
+    """Validate branches without changing normal order or disabled-option behavior."""
+    input_path = TYPE_UNION_INPUTS / f"{name}.json"
+    expected_file = f"type_union_constraints/{name}_{int(constraints)}.py"
+    if entry == "cli":
+        run_main_and_assert(
+            input_path=input_path,
+            output_path=output_file,
+            input_file_type="jsonschema",
+            extra_args=[
+                "--custom-file-header",
+                "# Type-specific union constraints",
+                *(["--field-constraints"] if constraints else []),
+            ],
+            assert_func=assert_file_content,
+            expected_file=expected_file,
+        )
+    elif entry == "api":
+        run_generate_file_and_assert(
+            input_path=input_path,
+            output_path=output_file,
+            input_file_type=InputFileType.JsonSchema,
+            custom_file_header="# Type-specific union constraints",
+            field_constraints=constraints,
+            assert_func=assert_file_content,
+            expected_file=expected_file,
+        )
+    schema = json.loads(input_path.read_text())
+    Draft202012Validator.check_schema(schema)
+    validator = Draft202012Validator(schema)
+    payloads = json.loads((TYPE_UNION_PAYLOADS / f"{name}.json").read_text())
+    assert_output(
+        json.dumps([validator.is_valid(value) for value in payloads], indent=2) + "\n",
+        TYPE_UNION_EXPECTED / f"{name}_native.txt",
+    )
+    runtime_path = TYPE_UNION_EXPECTED / (f"{name}_native.txt" if constraints else f"{name}_0_runtime.txt")
+    acceptance = json.loads(runtime_path.read_text())
+    context = (
+        nullcontext(
+            generate_dynamic_models(schema, config=GenerateConfig(field_constraints=constraints), cache_size=0)["Root"]
+        )
+        if entry == "dynamic"
+        else _generated_model(output_file, "type_union_constraints", "Root")
+    )
+    actual = []
+    with context as model:
+        for value, valid in zip(payloads, acceptance, strict=True):
+            if valid:
+                model.model_validate(value)
+                actual.append(True)
+            else:
+                with pytest.raises(ValidationError):
+                    model.model_validate(value)
+                actual.append(False)
+    assert_output(json.dumps(actual, indent=2) + "\n", runtime_path)
+
+
+@pytest.mark.parametrize(
+    "name",
+    ["string_length_field", "both_root", "array_minimum_field", "plain_field", "modeled_field", "numeric_min_field"],
+)
+@pytest.mark.parametrize(
+    "mode", ["annotated", "legacy_union", "strict", "msgspec", "pydantic_dataclass", "msgspec_plain"]
+)
+@pytest.mark.parametrize("entry", ["cli", "api"])
+def test_type_union_constraint_options(name: str, mode: str, entry: str, output_file: Path) -> None:
+    """Respect supported annotation, strictness, backend, and union-syntax choices."""
+    import msgspec
+    from pydantic import TypeAdapter
+
+    from datamodel_code_generator import DataModelType
+    from datamodel_code_generator.types import StrictTypes
+
+    options = {
+        "annotated": {"use_annotated": True},
+        "legacy_union": {"use_union_operator": False},
+        "strict": {"strict_types": [StrictTypes.str, StrictTypes.int]},
+        "msgspec": {"output_model_type": DataModelType.MsgspecStruct, "use_annotated": True},
+        "pydantic_dataclass": {"output_model_type": DataModelType.PydanticV2Dataclass},
+        "msgspec_plain": {"output_model_type": DataModelType.MsgspecStruct, "use_annotated": False},
+    }[mode]
+    arguments = {
+        "annotated": ["--use-annotated"],
+        "legacy_union": ["--no-use-union-operator"],
+        "strict": ["--strict-types", "str", "int"],
+        "msgspec": ["--output-model-type", "msgspec.Struct", "--use-annotated"],
+        "pydantic_dataclass": ["--output-model-type", "pydantic_v2.dataclass"],
+        "msgspec_plain": ["--output-model-type", "msgspec.Struct", "--no-use-annotated"],
+    }[mode]
+    input_path = TYPE_UNION_INPUTS / f"{name}.json"
+    expected_file = f"type_union_constraints/{name}_{mode}.py"
+    if entry == "cli":
+        run_main_and_assert(
+            input_path=input_path,
+            output_path=output_file,
+            input_file_type="jsonschema",
+            extra_args=["--custom-file-header", "# Type-specific union constraints", "--field-constraints", *arguments],
+            assert_func=assert_file_content,
+            expected_file=expected_file,
+        )
+    else:
+        run_generate_file_and_assert(
+            input_path=input_path,
+            output_path=output_file,
+            input_file_type=InputFileType.JsonSchema,
+            custom_file_header="# Type-specific union constraints",
+            field_constraints=True,
+            assert_func=assert_file_content,
+            expected_file=expected_file,
+            **options,
+        )
+    payloads = json.loads((TYPE_UNION_PAYLOADS / f"{name}.json").read_text())
+    expected_runtime = TYPE_UNION_EXPECTED / (
+        f"{name}_{mode}_runtime.txt" if mode == "msgspec_plain" else f"{name}_native.txt"
+    )
+    acceptance = json.loads(expected_runtime.read_text())
+    actual = []
+    with _generated_model(output_file, "type_union_options", "Root") as model:
+        validate = (
+            partial(msgspec.convert, type=model) if mode.startswith("msgspec") else TypeAdapter(model).validate_python
+        )
+        for value, valid in zip(payloads, acceptance, strict=True):
+            if valid:
+                validate(value)
+                actual.append(True)
+            else:
+                with pytest.raises((ValidationError, msgspec.ValidationError)):
+                    validate(value)
+                actual.append(False)
+    assert_output(json.dumps(actual, indent=2) + "\n", expected_runtime)
